Implement BartXmlParser abstract class
